Add useful wrapper methods for common use cases.
This adds the SendHello() and WriteRPC() methods. SendHello() will send
a pretty standard hello message as specified by the NETCONF protocol.
WriteRPC() takes an xml string and adds on the rpc tag, which is a
common use case for NETCONF. This also updates the example script to
show how xpath can be used to retrieve useful information out of the
data.